refactor: 잘못된 이름 수정과 북마크 옵셔널로 변경
