+++ frontend/src/components/Common/index.ts
@@ -8,7 +8,7 @@ export { default as SvgIcon } from './Svg/SvgIcon';
 export { default as TabMenu } from './TabMenu/TabMenu';
 export { default as TagList } from './TagList/TagList';
 export { default as Title } from './Title/Title';
-export { default as SectionTitle } from './SectionTitle/SectionTitleTitle';
+export { default as SectionTitle } from './SectionTitle/SectionTitle';
 export { default as ScrollButton } from './ScrollButton/ScrollButton';
 export { default as Input } from './Input/Input';
 export { default as ErrorBoundary } from './ErrorBoundary/ErrorBoundary';
